+++ lnbits/commands.py
@@ -12,7 +12,7 @@ from .core import migrations as core_migrations
 from .core.crud import get_dbversions, get_inactive_extensions
 from .core.helpers import migrate_extension_database, run_migration
 from .db import COCKROACH, POSTGRES, SQLITE
-from .extension_manger import get_valid_extensions
+from .extension_manager import get_valid_extensions
 from .helpers import get_css_vendored, get_js_vendored, url_for_vendored
